moved HelmetSeo for ebook from EbookLandingBanner to Ebook
fix unused-vars in EbookLandingBanner
